% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_georeferencia.R
\name{mod_georeferencia_server}
\alias{mod_georeferencia_server}
\title{georeferencia Server Functions}
\usage{
mod_georeferencia_server(id, dados)
}
\arguments{
\item{session}{}
}
\description{
georeferencia Server Functions
}
